+++ app/controllers/plans_controller.rb
@@ -27,12 +27,10 @@ class PlansController < ApplicationController
     @plan = Plan.find(params[:id])
 
     if @plan.update(plan_params)
-      flash.now.notice = notice('successfully.updated')
+      redirect_to edit_plan_path(@plan), notice: notice('successfully.updated')
     else
-      flash.now.alert = notice('could_not_be.updated')
+      redirect_to edit_plan_path(@plan), notice: notice('could_not_be.updated')
     end
-
-    render :edit
   end
 
   def destroy
@@ -60,9 +58,9 @@ class PlansController < ApplicationController
       :end_date,
       :user_id,
       activities_attributes: [
-        :id, :seconds_per_week, :_destroy,
-        { tag_associations_attributes: %i[id tag_type_id tag_id _destroy] }
-      ]
+        :id, :seconds_per_week, :_destroy, { tag_associations_attributes: %i[id tag_type_id tag_id _destroy] }
+      ],
+      signoffs_attributes: %i[id user_id _destroy]
     )
   end
 end